+++ annotation/src/commonMain/kotlin/com/cmgapps/LogTag.kt
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2026. Christian Grach <christian.grach@cmgapps.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package com.cmgapps
+
+@Retention(AnnotationRetention.SOURCE)
+@Target(
+    AnnotationTarget.CLASS,
+    AnnotationTarget.FUNCTION,
+)
+public annotation class LogTag(
+    val value: String = "",
+)