+++ compiler/compiler-plugin/src/main/kotlin/com/cmgapps/logtag/ir/LogTagIrGenerationExtension.kt
@@ -8,19 +8,16 @@ package com.cmgapps.logtag.ir
 
 import org.jetbrains.kotlin.backend.common.extensions.IrGenerationExtension
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
-import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
-import org.jetbrains.kotlin.ir.visitors.transformChildrenVoid
+import org.jetbrains.kotlin.ir.visitors.acceptChildrenVoid
 
-class LogTagIrGenerationExtension(
-    private val messageCollector: MessageCollector,
-) : IrGenerationExtension {
+class LogTagIrGenerationExtension : IrGenerationExtension {
     override fun generate(
         moduleFragment: IrModuleFragment,
         pluginContext: IrPluginContext,
     ) {
-        moduleFragment.transformChildrenVoid(
-            LogTagIrTransformer(context = pluginContext, messageCollector = messageCollector),
+        moduleFragment.acceptChildrenVoid(
+            LogTagIrGenerator(pluginContext),
         )
     }
 }