+++ compiler/compiler-plugin/src/main/kotlin/com/cmgapps/logtag/ir/LogTagIrGenerator.kt
@@ -10,34 +10,48 @@ import com.cmgapps.logtag.LOG_TAG_ANNOTATION_FQ_NAME
 import com.cmgapps.logtag.LOG_TAG_PROPERTY_NAME
 import com.cmgapps.logtag.LogTagPluginKey
 import org.jetbrains.kotlin.backend.common.extensions.IrPluginContext
-import org.jetbrains.kotlin.cli.common.messages.CompilerMessageSeverity
-import org.jetbrains.kotlin.cli.common.messages.MessageCollector
 import org.jetbrains.kotlin.descriptors.DescriptorVisibilities
-import org.jetbrains.kotlin.ir.IrStatement
+import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.builders.declarations.buildField
+import org.jetbrains.kotlin.ir.declarations.IrDeclaration
 import org.jetbrains.kotlin.ir.declarations.IrDeclarationOrigin
+import org.jetbrains.kotlin.ir.declarations.IrFile
+import org.jetbrains.kotlin.ir.declarations.IrModuleFragment
 import org.jetbrains.kotlin.ir.declarations.IrProperty
-import org.jetbrains.kotlin.ir.declarations.createExpressionBody
-import org.jetbrains.kotlin.ir.expressions.impl.IrConstImpl
-import org.jetbrains.kotlin.ir.util.classId
+import org.jetbrains.kotlin.ir.declarations.createBlockBody
+import org.jetbrains.kotlin.ir.expressions.impl.IrReturnImpl
 import org.jetbrains.kotlin.ir.util.findAnnotation
 import org.jetbrains.kotlin.ir.util.getAnnotationValueOrNull
 import org.jetbrains.kotlin.ir.util.parentAsClass
-import org.jetbrains.kotlin.ir.visitors.IrElementTransformerVoid
+import org.jetbrains.kotlin.ir.util.toIrConst
+import org.jetbrains.kotlin.ir.visitors.IrVisitorVoid
+import org.jetbrains.kotlin.ir.visitors.acceptChildrenVoid
 
-class LogTagIrTransformer(
+class LogTagIrGenerator(
     private val context: IrPluginContext,
-    private val messageCollector: MessageCollector,
-) : IrElementTransformerVoid() {
-    override fun visitProperty(declaration: IrProperty): IrStatement {
+) : IrVisitorVoid() {
+    override fun visitElement(element: IrElement) {
+        when (element) {
+            is IrDeclaration,
+            is IrFile,
+            is IrModuleFragment,
+            -> {
+                element.acceptChildrenVoid(this)
+            }
+
+            else -> {}
+        }
+    }
+
+    override fun visitProperty(declaration: IrProperty) {
         val origin = declaration.origin
 
         if (origin !is IrDeclarationOrigin.GeneratedByPlugin || origin.pluginKey != LogTagPluginKey) {
-            return super.visitProperty(declaration)
+            return
         }
 
         if (declaration.name != LOG_TAG_PROPERTY_NAME) {
-            return super.visitProperty(declaration)
+            return
         }
 
         val ownerClass = declaration.parentAsClass.let { if (it.isCompanion) it.parentAsClass else it }
@@ -52,16 +66,6 @@ class LogTagIrTransformer(
                 .orEmpty()
                 .ifBlank { ownerClass.name.asString().take(23) }
 
-        messageCollector.report(
-            CompilerMessageSeverity.INFO,
-            """
-            |Add property implementation
-            |   private val ${declaration.name} =  "$logTag"
-            |on
-            |   ${declaration.parentAsClass.classId?.asString()}
-            """.trimMargin(),
-        )
-
         declaration.backingField =
             context.irFactory
                 .buildField {
@@ -73,17 +77,28 @@ class LogTagIrTransformer(
                     parent = declaration.parent
                     initializer =
                         context.irFactory.createExpressionBody(
-                            IrConstImpl.string(
-                                startOffset,
-                                endOffset,
-                                context.irBuiltIns.stringType,
-                                logTag,
-                            ),
+                            startOffset = startOffset,
+                            endOffset = endOffset,
+                            logTag.toIrConst(context.irBuiltIns.stringType, startOffset, endOffset),
                         )
                 }
+        declaration.getter?.apply {
+            body =
+                context.irFactory.createBlockBody(
+                    startOffset,
+                    endOffset,
+                ) {
+                    statements +=
+                        IrReturnImpl(
+                            startOffset,
+                            endOffset,
+                            context.irBuiltIns.nothingType,
+                            symbol,
+                            logTag.toIrConst(context.irBuiltIns.stringType, startOffset, endOffset),
+                        )
+                }
+        }
         declaration.visibility = DescriptorVisibilities.PRIVATE
         declaration.annotations = emptyList()
-
-        return super.visitProperty(declaration)
     }
 }