+++ compiler/compiler-plugin/src/testFixtures/kotlin/com/cmgapps/logtag/compiler/GenerateTests.kt
@@ -6,6 +6,9 @@
 
 package com.cmgapps.logtag.compiler
 
+import com.cmgapps.logtag.compiler.runner.AbstractJsBoxOnlyTest
+import com.cmgapps.logtag.compiler.runner.AbstractJsBoxTest
+import com.cmgapps.logtag.compiler.runner.AbstractJsDiagnosticTest
 import com.cmgapps.logtag.compiler.runner.AbstractJvmBoxTest
 import com.cmgapps.logtag.compiler.runner.AbstractJvmDiagnosticTest
 import org.jetbrains.kotlin.generators.dsl.junit5.generateTestGroupSuiteWithJUnit5
@@ -17,8 +20,21 @@ fun main(args: Array<String>) {
                 model("diagnostics")
             }
 
+            testClass<AbstractJsDiagnosticTest> {
+                model("diagnostics")
+            }
+
             testClass<AbstractJvmBoxTest> {
                 model("box")
+                model("boxJvm")
+            }
+
+            testClass<AbstractJsBoxTest> {
+                model("box")
+            }
+
+            testClass<AbstractJsBoxOnlyTest> {
+                model("boxJs")
             }
         }
     }