+++ compiler/compiler-plugin/src/testFixtures/kotlin/com/cmgapps/logtag/compiler/service/PluginAnnotationsProvider.kt
@@ -8,6 +8,8 @@ package com.cmgapps.logtag.compiler.service
 
 import org.jetbrains.kotlin.cli.jvm.config.addJvmClasspathRoots
 import org.jetbrains.kotlin.config.CompilerConfiguration
+import org.jetbrains.kotlin.js.config.JSConfigurationKeys
+import org.jetbrains.kotlin.platform.isJs
 import org.jetbrains.kotlin.platform.jvm.isJvm
 import org.jetbrains.kotlin.test.builders.TestConfigurationBuilder
 import org.jetbrains.kotlin.test.model.TestModule
@@ -35,13 +37,13 @@ private class PluginAnnotationsProvider(
                 configuration.addJvmClasspathRoots(annotationsJvmRuntimeClasspath)
             }
 
-            // platform.isJs() -> {
-            //     val libraries = configuration.getList(JSConfigurationKeys.LIBRARIES)
-            //     configuration.put(
-            //         JSConfigurationKeys.LIBRARIES,
-            //         libraries + annotationsJsRuntimeClasspath.map { it.absolutePath },
-            //     )
-            // }
+            platform.isJs() -> {
+                val libraries = configuration.getList(JSConfigurationKeys.LIBRARIES)
+                configuration.put(
+                    JSConfigurationKeys.LIBRARIES,
+                    libraries + annotationsJsRuntimeClasspath.map { it.absolutePath },
+                )
+            }
         }
     }
 }
@@ -53,15 +55,14 @@ private class PluginAnnotationsClasspathProvider(
         val targetPlatform = module.targetPlatform(testServices)
         return when {
             targetPlatform.isJvm() -> annotationsJvmRuntimeClasspath
-
-            // targetPlatform.isJs() -> annotationsJsRuntimeClasspath
+            targetPlatform.isJs() -> annotationsJsRuntimeClasspath
             else -> emptyList()
         }
     }
 }
 
 private val annotationsJvmRuntimeClasspath = classpathFiles("annotationsRuntime.jvm.classpath")
-// private val annotationsJsRuntimeClasspath = classpathFiles("annotationsRuntime.js.classpath")
+private val annotationsJsRuntimeClasspath = classpathFiles("annotationsRuntime.js.classpath")
 
 private fun classpathFiles(property: String): List<File> {
     val property =