+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Changed
 
+- Annotation library `com.cmgapps.logtag:annotation` is not Kotlin Multiplatform compatible
+
 ### Deprecated
 
 - LogTag's KAPT will be removed in the next major release; change to the KSP version